Trying to hunt down a bug

+++ app/controllers/log_file_controller.rb
@@ -26,33 +26,35 @@ class LogFileController < ApplicationController
   # 
   # Take the given file and save it to disc.
   def save
-    uploaded_io = params[:log_file]
+    if current_user.admin? == false
+      uploaded_io = params[:log_file]
 
-    saved_filename = Rails.root.join('public', 'uploads', "#{Time.new.strftime("%Y-%m-%d-%H%M%S")}_#{uploaded_io.original_filename}")
+      saved_filename = Rails.root.join('public', 'uploads', "#{Time.new.strftime("%Y-%m-%d-%H%M%S")}_#{uploaded_io.original_filename}")
 
-    File.open(saved_filename, 'wb') do |file|
-      file.write(uploaded_io.read)
-    end
-    
-    line_number = 0
-    file_suffix = 1
-    
-    IO.foreach(saved_filename) do |line|
-      if(line_number % 200 == 0)
-        file_suffix = file_suffix + 1
+      File.open(saved_filename, 'wb') do |file|
+        file.write(uploaded_io.read)
       end
+    
+      line_number = 0
+      file_suffix = 1
+    
+      IO.foreach(saved_filename) do |line|
+        if(line_number % 200 == 0)
+          file_suffix = file_suffix + 1
+        end
 
-      File.open(Rails.root.join('public', 'uploads', "#{saved_filename}.split#{file_suffix}.log"), "a+") do |split_file|
-        split_file.write(line)
-      end
+        File.open(Rails.root.join('public', 'uploads', "#{saved_filename}.split#{file_suffix}.log"), "a+") do |split_file|
+          split_file.write(line)
+        end
 
-      line_number = line_number + 1
+        line_number = line_number + 1
       
+      end
+    
+      File.delete(saved_filename)
+      processor = LogFileProcessor.new
+    
+      processor.process_log_files()
     end
-    
-    File.delete(saved_filename)
-    processor = LogFileProcessor.new
-    
-    processor.process_log_files()
   end
 end
